+++ android/src/main/kotlin/com/kino/video_360/Video360View.kt
@@ -3,29 +3,27 @@ package com.kino.video_360
 import android.app.Activity
 import android.app.Application
 import android.content.Context
-import android.content.Intent
-import android.graphics.Color
 import android.os.Bundle
 import android.util.Log
 import android.view.View
 import android.view.ViewGroup
 import android.widget.FrameLayout
-import com.kino.video_360.Video360UIView
 import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.platform.PlatformView
 import io.flutter.view.TextureRegistry
-import java.util.*
-import kotlin.collections.HashMap
 
-class Video360View(context: Context, messenger: BinaryMessenger, id: Int, textureRegistry: TextureRegistry)
+class Video360View(context: Context,
+                   messenger: BinaryMessenger,
+                   id: Int,
+                   textureRegistry: TextureRegistry)
     : PlatformView, MethodChannel.MethodCallHandler {
 
-    private val TAG: String = Video360View::class.java.simpleName
+    private val tag: String = Video360View::class.java.simpleName
 
     private val methodChannel: MethodChannel = MethodChannel(messenger, "kino_video_360_$id")
-    lateinit var activityLifecycleCallbacks: Application.ActivityLifecycleCallbacks
+    private lateinit var activityLifecycleCallbacks: Application.ActivityLifecycleCallbacks
 
     private var videoView: Video360UIView
 
@@ -45,27 +43,21 @@ class Video360View(context: Context, messenger: BinaryMessenger, id: Int, textur
         when (call.method) {
             "init" -> {
                 val url: String? = call.argument("url")
-                val isAutoPlay: Boolean? = call.argument("isAutoPlay")
-                val isRepeat: Boolean? = call.argument("isRepeat")
-                url?.let { vUrl ->
-                    isAutoPlay?.let { autoPlay ->
-                        isRepeat?.let { repeat ->
-                            videoView.initializePlayer(vUrl, autoPlay, repeat)
-//                            videoView.context.startActivity(Intent(videoView.context, VRActivity::class.java))
-                        }
-                    }
+                val isRepeat: Boolean = call.argument("isRepeat") ?: false
+                url?.let {
+                    videoView.initializePlayer(it, false, isRepeat)
                 }
             }
             "resume" -> {
-                Log.i(TAG, "resume")
+                Log.d(tag, "resume")
                 onResume()
             }
             "pause" -> {
-                Log.i(TAG, "pause")
+                Log.d(tag, "pause")
                 onPause()
             }
             "dispose" -> {
-                Log.i(TAG, "dispose")
+                Log.d(tag, "dispose")
                 dispose()
             }
             "play" -> {
@@ -109,26 +101,26 @@ class Video360View(context: Context, messenger: BinaryMessenger, id: Int, textur
     private fun setupLifeCycle(context: Context) {
         activityLifecycleCallbacks = object : Application.ActivityLifecycleCallbacks {
             override fun onActivityCreated(activity: Activity, savedInstanceState: Bundle?) {
-                Log.i(TAG, "onActivityCreated")
+                Log.d(tag, "onActivityCreated")
             }
 
             override fun onActivityStarted(activity: Activity) {
-                Log.i(TAG, "onActivityStarted")
+                Log.d(tag, "onActivityStarted")
                 onStart()
             }
 
             override fun onActivityResumed(activity: Activity) {
-                Log.i(TAG, "onActivityResumed")
+                Log.d(tag, "onActivityResumed")
                 onResume()
             }
 
             override fun onActivityPaused(activity: Activity) {
-                Log.i(TAG, "onActivityPaused")
+                Log.d(tag, "onActivityPaused")
                 onPause()
             }
 
             override fun onActivityStopped(activity: Activity) {
-                Log.i(TAG, "onActivityStopped")
+                Log.d(tag, "onActivityStopped")
                 onStop()
             }
 
@@ -136,7 +128,7 @@ class Video360View(context: Context, messenger: BinaryMessenger, id: Int, textur
             }
 
             override fun onActivityDestroyed(activity: Activity) {
-                Log.i(TAG, "onActivityDestroyed")
+                Log.d(tag, "onActivityDestroyed")
                 onDestroy()
             }
         }