+++ example/lib/main.dart
@@ -1,14 +1,20 @@
 import 'package:flutter/material.dart';
 import 'package:video_360/video_360.dart';
 
-void main() {
+Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  runApp(MaterialApp(home: MyApp()));
+  runApp(const MaterialApp(
+    home: MyApp(),
+  ));
 }
 
 class MyApp extends StatefulWidget {
+  const MyApp({
+    super.key,
+  });
+
   @override
-  _MyAppState createState() => _MyAppState();
+  State<MyApp> createState() => _MyAppState();
 }
 
 class _MyAppState extends State<MyApp> {
@@ -24,11 +30,6 @@ class _MyAppState extends State<MyApp> {
 
   @override
   Widget build(BuildContext context) {
-    var statusBar = MediaQuery.of(context).padding.top;
-
-    var width = MediaQuery.of(context).size.width;
-    var height = MediaQuery.of(context).size.height;
-
     return Scaffold(
       appBar: AppBar(
         title: const Text('Video 360 Plugin example app'),
@@ -36,13 +37,12 @@ class _MyAppState extends State<MyApp> {
       body: Stack(
         children: [
           Center(
-            child: Container(
-              width: width,
-              height: height,
+            child: SizedBox(
+              width: MediaQuery.of(context).size.width,
+              height: MediaQuery.of(context).size.height,
               child: Video360View(
+                url: 'https://bitmovin-a.akamaihd.net/content/playhouse-vr/m3u8s/105560.m3u8',
                 onVideo360ViewCreated: _onVideo360ViewCreated,
-                url:
-                    'https://bitmovin-a.akamaihd.net/content/playhouse-vr/m3u8s/105560.m3u8',
                 onPlayInfo: (Video360PlayInfo info) {
                   setState(() {
                     durationText = info.duration.toString();
@@ -62,28 +62,28 @@ class _MyAppState extends State<MyApp> {
                       controller?.play();
                     },
                     color: Colors.grey[100],
-                    child: Text('Play'),
+                    child: const Text('Play'),
                   ),
                   MaterialButton(
                     onPressed: () {
                       controller?.stop();
                     },
                     color: Colors.grey[100],
-                    child: Text('Stop'),
+                    child: const Text('Stop'),
                   ),
                   MaterialButton(
                     onPressed: () {
                       controller?.reset();
                     },
                     color: Colors.grey[100],
-                    child: Text('Reset'),
+                    child: const Text('Reset'),
                   ),
                   MaterialButton(
                     onPressed: () {
                       controller?.jumpTo(80000);
                     },
                     color: Colors.grey[100],
-                    child: Text('1:20'),
+                    child: const Text('1:20'),
                   ),
                 ],
               ),
@@ -95,20 +95,20 @@ class _MyAppState extends State<MyApp> {
                       controller?.seekTo(-2000);
                     },
                     color: Colors.grey[100],
-                    child: Text('<<'),
+                    child: const Text('<<'),
                   ),
                   MaterialButton(
                     onPressed: () {
                       controller?.seekTo(2000);
                     },
                     color: Colors.grey[100],
-                    child: Text('>>'),
+                    child: const Text('>>'),
                   ),
                   Flexible(
                     child: MaterialButton(
                       onPressed: () {},
                       color: Colors.grey[100],
-                      child: Text(durationText + ' / ' + totalText),
+                      child: Text('$durationText / $totalText'),
                     ),
                   ),
                 ],
@@ -120,7 +120,8 @@ class _MyAppState extends State<MyApp> {
     );
   }
 
-  _onVideo360ViewCreated(Video360Controller? controller) {
+  void _onVideo360ViewCreated(Video360Controller controller) {
     this.controller = controller;
+    this.controller?.play();
   }
 }