+++ lib/src/video360_play_info.dart
@@ -1,8 +1,0 @@
-class Video360PlayInfo {
-  int duration = 0;
-  int total = 0;
-  bool isPlaying = false;
-
-  Video360PlayInfo(
-      {required this.duration, required this.total, required this.isPlaying});
-}