+++ lib/src/video_360_play_info.dart
@@ -1,0 +1,12 @@
+class Video360PlayInfo {
+
+  Video360PlayInfo({
+    required this.duration,
+    required this.total,
+    required this.isPlaying,
+  });
+
+  final int duration;
+  final int total;
+  final bool isPlaying;
+}