+++ lib/video_360.dart
@@ -1,7 +1,5 @@
 library video_360;
 
-export 'package:video_360/src/video360_view.dart';
-export 'package:video_360/src/video360_android_view.dart';
-export 'package:video_360/src/video360_ios_view.dart';
-export 'package:video_360/src/video360_controller.dart';
-export 'package:video_360/src/video360_play_info.dart';
+export 'package:video_360/src/video_360_play_info.dart';
+export 'package:video_360/src/video_360_controller.dart';
+export 'package:video_360/src/view/video_360_view.dart';